Changes in support of using a database directory local to the db instance, because the sshfs-mounted directory was freezing on load. I have since found this helpful information about this problem:
    https://discourse.ubuntu.com/t/mount-i-o-slow-how-to-improve-performance/18479/5

I don't like this solution as much, but it will do for now. Next, I'll look into using
a native KVM virtio-based mount to overcome this issue.

NOTE: this change also includes some performance tweaks for the db: the 'paths' materialized
view and switching the log dir to a different disk than the db.